// apps/web/src/api/model/footnoteSentenceResponse.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Scholia API
 * OpenAPI spec version: 0.1.0
 */

export interface FootnoteSentenceResponse {
    html: string;
    id: string;
    /** @nullable */
    original_html?: string | null;
    /** @nullable */
    original_text?: string | null;
    position: number;
    /** @nullable */
    sentence_number?: number | null;
    text: string;
}